Notify users that update is downloading Add a "Downloading I2P Browser update" item which appears in the hamburger (app) menu while the update service is downloading a MAR file. Before this change, the browser did not indicate to the user that an update was in progress, which is especially confusing in I2P Browser because downloads often take some time. If the user clicks on the new menu item, the about dialog is opened to allow the user to see download progress.
As part of this fix, the update service was changed to always show
update-related messages in the hamburger menu, even if the update
was started in the foreground via the about dialog or via the
"Check for I2P Browser Update" toolbar menu item. This change is
consistent with the I2P Browser goal of making sure users are
informed about the update process.